+++ backend/src/main/java/toonpick/app/domain/toon_collection/ToonCollection.java
@@ -1,4 +1,4 @@
-package toonpick.app.domain.collection;
+package toonpick.app.domain.toon_collection;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
