+++ backend/src/main/java/toonpick/app/repository/BookmarkToonsRepository.java
@@ -1,9 +1,10 @@
-package toonpick.app.domain.WatchingToons;
+package toonpick.app.repository;
 
 import io.lettuce.core.dynamic.annotation.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+import toonpick.app.domain.bookmark_toons.BookmarkToons;
 import toonpick.app.domain.member.Member;
 import toonpick.app.domain.webtoon.Webtoon;
 
@@ -11,11 +12,11 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface WatchingToonsRepository extends JpaRepository<WatchingToons, Long> {
+public interface BookmarkToonsRepository extends JpaRepository<BookmarkToons, Long> {
 
-    Optional<WatchingToons> findByMemberAndWebtoon(Member member, Webtoon webtoon);
+    Optional<BookmarkToons> findByMemberAndWebtoon(Member member, Webtoon webtoon);
 
-    @Query("SELECT w.webtoon FROM WatchingToons w WHERE w.member = :member")
+    @Query("SELECT b.webtoon FROM BookmarkToons b WHERE b.member = :member")
     List<Webtoon> findWebtoonsByMember(@Param("member") Member member);
 
 }
