+++ backend/src/main/java/toonpick/app/repository/ToonCollectionRepository.java
@@ -1,7 +1,8 @@
-package toonpick.app.domain.collection;
+package toonpick.app.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import toonpick.app.domain.toon_collection.ToonCollection;
 import toonpick.app.domain.member.Member;
 
 import java.util.List;
